requirements.txt
langchain
langchain-openai
langchain-community
langchain-qdrant
langgraph
qdrant-client
streamlit
python-dotenv
requests
pypdf
tiktoken
langsmith
pytest
reportlab
